# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Bus Ticket App
  description: |-
    REST API for managing bus ticket reservations, trips, companies, vehicles and payments.
    Built with Spring Boot and integrated with Stripe for payment processing.

    Some useful links:
    - [Bus Ticket App repository](https://github.com/texksi/Bus-Ticket-App)
  contact:
    email: teamikic18@gmail.com
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
...
paths:
  /api/korisnici:
    get:
      tags:
        - Korisnik
      summary: Get all users
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KorisnikResponse'
    post:
      tags:
        - Korisnik
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KorisnikRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KorisnikResponse'
        '409':
          description: User already exists
  /api/korisnici/{id}:
    get:
      tags:
        - Korisnik
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KorisnikResponse'
        '404':
          description: User not found
    put:
      tags:
        - Korisnik
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KorisnikRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KorisnikResponse'
        '404':
          description: User not found
    delete:
      tags:
        - Korisnik
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
  /api/korisnici/{id}/rezervacije:
    get:
      tags:
        - Korisnik
      summary: Get all reservations for user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RezervacijaResponse'
  /api/korisnici/{id}/ocene:
    get:
      tags:
        - Korisnik
      summary: Get all ratings for user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OcenaResponse'
  /api/rezervacije:
    get:
      tags:
        - Rezervacija
      summary: Get all reservations
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RezervacijaResponse'
    post:
      tags:
        - Rezervacija
      summary: Create new reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RezervacijaRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RezervacijaResponse'
  /api/rezervacije/{id}:
    get:
      tags:
        - Rezervacija
      summary: Get reservation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RezervacijaResponse'
        '404':
          description: Reservation not found
    put:
      tags:
        - Rezervacija
      summary: Update reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RezervacijaRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RezervacijaResponse'
    delete:
      tags:
        - Rezervacija
      summary: Delete reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
  /api/rezervacije/{id}/karte:
    get:
      tags:
        - Rezervacija
      summary: Get all tickets for reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KartaResponse'
...
  /api/kompanije:
    get:
      tags:
        - Kompanija
      summary: Get all companies
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KompanijaResponse'
    post:
      tags:
        - Kompanija
      summary: Create new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KompanijaRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KompanijaResponse'
  /api/kompanije/{id}:
    get:
      tags:
        - Kompanija
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KompanijaResponse'
        '404':
          description: Company not found
    put:
      tags:
        - Kompanija
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KompanijaRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KompanijaResponse'
    delete:
      tags:
        - Kompanija
      summary: Delete company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
  /api/kompanije/{id}/putovanja:
    get:
      tags:
        - Kompanija
      summary: Get all trips for company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PutovanjeResponse'
  /api/kompanije/{id}/vozila:
    get:
      tags:
        - Kompanija
      summary: Get all vehicles for company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoziloResponse'
...
components:
  schemas:
    KorisnikRequest:
      type: object
      properties:
        ime:
          type: string
          example: "Marko"
        prezime:
          type: string
          example: "Markovic"
        email:
          type: string
          format: email
          example: "marko@email.com"
        username:
          type: string
          example: "marko123"
        password:
          type: string
          example: "lozinka123"
        role:
          type: string
          enum:
            - ADMIN
            - USER
    KorisnikResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ime:
          type: string
        prezime:
          type: string
        email:
          type: string
        username:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
    RezervacijaRequest:
      type: object
      properties:
        nacinPlacanja:
          type: string
          example: "Kartica"
        status:
          type: string
          example: "pending"
        ukupanIznos:
          type: number
          example: 1500
        korisnikId:
          type: integer
          format: int64
          example: 1
    RezervacijaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        datumKreiranja:
          type: string
          format: date-time
        ukupanIznos:
          type: number
        nacinPlacanja:
          type: string
        status:
          type: string
        korisnikId:
          type: integer
          format: int64
...
    KartaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        brojSedista:
          type: string
        osnovnaCena:
          type: number
        datumIzdavanja:
          type: string
          format: date-time
        tip:
          type: string
        rezervacijaId:
          type: integer
          format: int64
        putovanjeId:
          type: integer
          format: int64
...
    PutovanjeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        polaziste:
          type: string
        odrediste:
          type: string
        vremePolaska:
          type: string
          format: date-time
        vremeDolaska:
          type: string
          format: date-time
        osnovnaCena:
          type: number
        kompanijaId:
          type: integer
          format: int64
        voziloId:
          type: integer
          format: int64
...
    OcenaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        komentar:
          type: string
        ocena:
          type: integer
        korisnikId:
          type: integer
          format: int64
        putovanjeId:
          type: integer
          format: int64
...
    KompanijaRequest:
      type: object
      properties:
        naziv:
          type: string
          example: "BusLine"
        kontakt:
          type: string
          example: "busline@email.com"
    KompanijaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        naziv:
          type: string
        kontakt:
          type: string
...
    VoziloResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        registracija:
          type: string
        kapacitet:
          type: integer
        brojRedova:
          type: integer
        brojKolona:
          type: integer
        kompanijaId:
          type: integer
          format: int64